+++ frontend/app/lib/analytics.tsx
